调整

+++ stone/common.py
@@ -2,14 +2,15 @@
 #作者：琴弦上的宇宙
 #时间：2021年10月22日
 
-import cv2
+#import cv2
 import pyautogui
 import time
 import random
+from myGui import *
 
 def moveAndClick(x,y):
     if x>10 and x<=1910 and y>10 and y<=1070:
-        pyautogui.moveTo(x+randint(-5, 5),y+randint(-5, 5))
+        pyautogui.moveTo(x+random.randint(-5, 5),y+random.randint(-5, 5))
         pyautogui.click(clicks=1)
         pyautogui.moveTo(650,20)
         pyautogui.click(clicks=1)
@@ -19,10 +20,10 @@ def moveAndClick(x,y):
 
 def drag(x0,y0,x1,y1):
     if x0>10 and x0<=1910 and y0>10 and y0<=1070 and x1>10 and x1<=1910 and y1>10 and y1<=1070:
-        pyautogui.mouseDown(x0+randint(-5,5),y0+randint(-5,5))
-        pyautogui.moveTo(x1+randint(-5,5),y1+randint(-5,5))
+        pyautogui.mouseDown(x0+random.randint(-5,5),y0+random.randint(-5,5))
+        pyautogui.moveTo(x1+random.randint(-5,5),y1+random.randint(-5,5))
         time.sleep(0.5)
-        pyautogui.mouseUp(x1+randint(-5,5),y1+randint(-5,5))
+        pyautogui.mouseUp(x1+random.randint(-5,5),y1+random.randint(-5,5))
         pyautogui.moveTo(650,20)
         pyautogui.click(clicks=1)
     else:
@@ -30,35 +31,28 @@ def drag(x0,y0,x1,y1):
         pyautogui.click(clicks=1)
 
 def bFindInBackground(background,png):
-    result=cv2.matchTemplate(background,png,cv2.TM_CCOEFF_NORMED)
-    minVal,maxVal,minLoc,maxLoc=cv2.minMaxLoc(result)
-    if maxVal>0.8:
-        x=maxLoc[0]
-        y=maxLoc[1]
-        w,h=png.shape[::-1]
-        if 0:
-            left_top=maxLoc
-            right_bottom=(left_top[0]+w,left_top[1]+h)
-            img=cv2.imread("resource/background.png",1)
-            cv2.rectangle(img,left_top,right_bottom,(0,0,255),2)
-            img=cv2.resize(img,(0,0),fx=0.5,fy=0.5,interpolation=cv2.INTER_NEAREST)
-            print('maxVal='),
-            print(maxVal)
-            cv2.imshow("processed",img)
-            cv2.waitKey(0)
-            cv2.destroyAllWindows()
-        return True,x,y,w,h
-    else:
-        return False,0,0,0,0
-
-class runStatus(object):
-    _bRunning=True
-    _runTime=int(time.time())
-    def __init__(self):pass
-    def bRunning(self):return self._bRunning
-    def setRunning(self,result):self._bRunning=result
-    def getRunTime(self):return runStatus._runTime
-    def setRunTime(self):runStatus._runTime=int(time.time())
+    if 0:
+        result=cv2.matchTemplate(background,png,cv2.TM_CCOEFF_NORMED)
+        minVal,maxVal,minLoc,maxLoc=cv2.minMaxLoc(result)
+        if maxVal>0.8:
+            x=maxLoc[0]
+            y=maxLoc[1]
+            w,h=png.shape[::-1]
+            if MyGui.gDebug:
+                left_top=maxLoc
+                right_bottom=(left_top[0]+w,left_top[1]+h)
+                img=cv2.imread("resource/background.png",1)
+                cv2.rectangle(img,left_top,right_bottom,(0,0,255),2)
+                img=cv2.resize(img,(0,0),fx=0.5,fy=0.5,interpolation=cv2.INTER_NEAREST)
+                print('maxVal='),
+                print(maxVal)
+                cv2.imshow("processed",img)
+                cv2.waitKey(0)
+                cv2.destroyAllWindows()
+            return True,x,y,w,h
+        else:
+            return False,0,0,0,0
+    return False,0,0,0,0
 
 class myPng(object):
     def __init__(self,path,name):
@@ -67,7 +61,8 @@ class myPng(object):
         if path!='' and name!='':
             p=path+'/'+str(name)
             p.replace("'","\"")
-            self.png=cv2.imread(p,cv2.IMREAD_GRAYSCALE)
+            #self.png=cv2.imread(p,cv2.IMREAD_GRAYSCALE)
+            self.png=None
 
 class myScene(object):
